requirement.txt
date-extractor==3.9.1
datefinder==0.7.0
Django==2.1.7
fpdf==1.7.2
nltk==3.4
numpy==1.16.1
pandas==0.24.1
Pillow==5.4.1
pyap==0.2.0
pytesseract==0.2.6
python-dateutil==2.8.0
pytz==2018.9
regex==2019.2.20
singledispatch==3.4.0.3
six==1.12.0
xlrd==1.2.0
